refactor: `JPA` `UserRepository` 폴더 이동
